+++ test/functional/authentication_tests.js
@@ -289,7 +289,7 @@ exports['Unordered bulk operation should fail correctly when not authenticated']
 var replSetManager;
 
 var setUp = function(configuration, options, callback) {
-  var ReplSetManager = require('mongodb-core').ReplSetManager
+  var ReplSetManager = require('mongodb-tools').ReplSetManager
     , Db = configuration.require.Db
     , Server = configuration.require.Server
     , MongoClient = configuration.require.MongoClient;
@@ -908,12 +908,10 @@ exports.shouldCorrectlyBringReplicasetStepDownPrimaryAndStillReadFromSecondary =
       db.open(function(err, db_p) {});
       db.on('all', function(err, db_p) {
         test.ok(db_p != null);
-        console.log("--------------------------------------------- 0")
         db_p.admin().addUser("me", "secret", {w:4, wtimeout:25000}, function runWhatever(err, result) {
           // Just set auths for the manager to handle it correctly
           replSetManager.setCredentials("mongocr", "admin", "me", "secret");
 
-        console.log("--------------------------------------------- 1")
           db_p.admin().authenticate("me", "secret", function(err, result) {
             test.equal(null, err);
 
@@ -924,10 +922,7 @@ exports.shouldCorrectlyBringReplicasetStepDownPrimaryAndStillReadFromSecondary =
                 test.equal(null, err);                
                 test.ok(result != null);
 
-        console.log("--------------------------------------------- 2")
                   db.serverConfig.on('joined', function(t) {
-        console.log("--------------------------------------------- 3")
-        console.dir(t)
                     if(t == 'primary') {
                       var counter = 1000;
                       var errors = 0;
@@ -954,8 +949,6 @@ exports.shouldCorrectlyBringReplicasetStepDownPrimaryAndStillReadFromSecondary =
                     }
                   });
                   db.serverConfig.on('left', function(t) {
-        console.log("--------------------------------------------- 4")
-        console.dir(t)
                   });
                 // Step down the primary
                 replSetManager.stepDown({force:true}, function(err, result) {
@@ -1206,7 +1199,7 @@ exports.shouldCorrectlyAuthAgainstNormalDbUsingMongoClient = {
 var shardedManager;
 
 var setUpSharded = function(configuration, options, callback) {
-  var ShardingManager = require('mongodb-core').ShardingManager
+  var ShardingManager = require('mongodb-tools').ShardingManager
     , Db = configuration.require.Db
     , Server = configuration.require.Server
     , MongoClient = configuration.require.MongoClient